Rename: 대시보드 컴포넌트 폴더 생성
